+++ km3pipe/cmd.py
@@ -6,12 +6,12 @@ KM3Pipe command line utility.
 Usage:
     km3pipe test
     km3pipe update [GIT_BRANCH]
-    km3pipe detx DET_ID [-m] [-t T0_SET] [-c CALIBR_ID]
+    km3pipe detx DET_ID [-m] [-t T0_SET] [-c CALIBR_ID] [-o OUTFILE]
     km3pipe detectors [-s REGEX] [--temporary]
     km3pipe runtable [-n RUNS] [-s REGEX] [--temporary] DET_ID
     km3pipe runinfo [--temporary] DET_ID RUN
     km3pipe rundetsn [--temporary] RUN DETECTOR
-    km3pipe retrieve DET_ID RUN
+    km3pipe retrieve DET_ID RUN [-o OUTFILE]
     km3pipe (-h | --help)
     km3pipe --version
 
@@ -19,6 +19,7 @@ Options:
     -h --help           Show this screen.
     -m                  Get the MC detector file (flips the sign of DET_ID).
     -c CALIBR_ID        Geometrical calibration ID (eg. A01466417)
+    -o OUTFILE          Output filename.
     -t T0_SET           Time calibration ID (eg. A01466431)
     -n EVENTS/RUNS      Number of events/runs.
     -s REGEX            Regular expression to filter the runsetup name/id.
@@ -110,21 +111,23 @@ def update_km3pipe(git_branch=''):
               .format(git_branch))
 
 
-def retrieve(run_id, det_id):
+def retrieve(run_id, det_id, outfile=None):
     """Retrieve run from iRODS for a given detector (O)ID"""
     try:
         det_id = int(det_id)
     except ValueError:
         pass
     path = irods_filepath(det_id, run_id)
-    os.system("iget -Pv {0}".format(path))
+    suffix = '' if outfile is None else outfile
+    os.system("iget -Pv {0} {1}".format(path, suffix))
 
 
-def detx(det_id, calibration='', t0set=''):
+def detx(det_id, calibration='', t0set='', filename=None):
     now = datetime.now()
-    filename = "KM3NeT_{0}{1:08d}_{2}.detx" \
-               .format('-' if det_id < 0 else '', abs(det_id),
-                       now.strftime("%d%m%Y"))
+    if filename is None:
+        filename = "KM3NeT_{0}{1:08d}_{2}.detx" \
+                   .format('-' if det_id < 0 else '', abs(det_id),
+                           now.strftime("%d%m%Y"))
     det = Detector(det_id=det_id, t0set=t0set, calibration=calibration)
     if det.n_doms > 0:
         det.write(filename)
@@ -174,13 +177,14 @@ def main():
         rundetsn(int(args['RUN']), args['DETECTOR'], temporary=args["--temporary"])
 
     if args['retrieve']:
-        retrieve(int(args['RUN']), args['DET_ID'])
+        retrieve(int(args['RUN']), args['DET_ID'], args['-o'])
 
     if args['detx']:
         t0set = args['-t']
         calibration = args['-c']
+        outfile = args['-o']
         det_id = int(('-' if args['-m'] else '') + args['DET_ID'])
-        detx(det_id, calibration, t0set)
+        detx(det_id, calibration, t0set, outfile)
 
     if args['detectors']:
         detectors(regex=args['-s'], temporary=args["--temporary"])